requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
safetensors==0.4.3
accelerate==0.31.0
transformers==4.41.2
diffusers==0.29.2
huggingface_hub==0.24.5
